refactor: separar el requirement
